:knife: :dragon_face: :dragon:  Major Update: 8080 SLY lexer and parser. :dragon: :dragon_face: :knife:
The - in MVI A,'-' is being picked up by math_expression instead of immediate_data.
This causes a shift in parsing, which is labeling incorrect lines as syntax errors.
The lexer is technically the problem, not the parser for this specific problem.

Almost slain the lexing and parsing dragons. The 8080_fp16.ASM almost lexes and parses in it's entirety. This is a big step, as then I can focus on actually translating the AST into ARMv6-M ISA.