:finnadie: PLY now handles Labels. Adds all instructions in a label to a list :finnadie:
Problem: PLY_I8080.py adds a label and instruction twice.
['LABEL1', ('MOV', ('A', 'B')), (['LABEL1', ('MOV', ('A', 'B'))],), ('ADD', ('M', '10H')), ('LXI', ('H', ('F16_1', '+', '1')))]
['CHR', ('ADD', ('A', ('-', '1'))), (['CHR', ('ADD', ('A', ('-', '1')))],)]

You can see it's adding the label and it's first instruction twice.
This was happening because the label and its corresponding first instruction are added to the AST twice:
    once as a label and then again when the instruction itself is parsed.

This will need a rethink.

If a label is encountered, it's output encapsulates the first instruction by it's label like so
['LABEL1', ('MOV', ('A', 'B')), ('LABEL1',), ('ADD', ('M', '10H')), ('LXI', ('H', ('F16_1', '+', '1')))]
['CHR', ('ADD', ('A', ('-', '1'))), ('CHR',)]
Need to fix the first instruction label encapsulation, but it's a ton better than before.

self.ast = [] in parser is used and the file is now being parsed into it.
The handling of this is quite bad, but it works for now.