:finnadie: C/C++ Pico BIOS/GIOS and HW-Agnostic BDOS/GDOS :finnadie:
C/C++ example of a Pico BIOS/GIOS and HW-Agnostic BDOS/GDOS.
The entire idea of this is that you could rewrite bios and gios for different hardware, while getting bdos and gdos work on the new hardware. No, none of it has been compiled or tested.
I'm sure a lot needs to be fixed/changed to align with bdos/gdos and to compile correctly.

More reference.

Refactor pico-asm-main folder